Split Entity
